requirements.txt
flask
flask-login
msal
supabase
google-cloud-aiplatform
python-dotenv
gunicorn